Move ilsvrc2012 challenge to a separate folder
