+++ dbcollection/datasets/imagenet/ilsvrc2012/classification.py
